// src/enums/categories.ts
export enum CATEGORIES {
  ARTS_AND_LITERATURE = 'arts_and_literature',
  FILM_AND_TV = 'film_and_tv',
  FOOD_AND_DRINK = 'food_and_drink',
  GENERAL_KNOWLEDGE = 'general_knowledge',
  GEOGRAPHY = 'geography',
  HISTORY = 'history',
  MUSIC = 'music',
  SCIENCE = 'science',
  SOCIETY_AND_CULTURE = 'society_and_culture',
  SPORT_AND_LEISURE = 'sport_and_leisure'
}